Convert Level rendering to use TileSets
Added a basic tileset (thanks to yd @ http://opengameart.org/content/tiles-sands)
that can be used to make more complex levels. We'll need to eventually make
a 1to1 mapping of the worldimage to tileset index so the code doesn't have
to hardcard tile values. A moved to actual level files (instead of images)
will solve this problem.